// desktop/src/open-api/model/reportDetail.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReportDetail { 
    mode: ReportDetail.ModeEnum;
    /**
     * Dimension field key to aggregate leaf rows by (aggregate mode)
     */
    by?: string;
}
export namespace ReportDetail {
    export type ModeEnum = 'records' | 'aggregate';
    export const ModeEnum = {
        Records: 'records' as ModeEnum,
        Aggregate: 'aggregate' as ModeEnum
    };
}